refactor: register field() as a field specifier
The custom `field` wrapper around `equinox.field` is not in any
PEP 681 `field_specifiers` tuple, so type checkers (e.g. ty) read
`x: T = field(...)` as a defaulted field and wrongly flag every
later required field as a "required field after a defaulted one"
(ty's `dataclass-field-order`).

Add `bartz._jaxext.Module`, an `equinox.Module` whose metaclass
carries `@dataclass_transform` listing `dataclasses.field`,
`equinox.field` and this `field` as `field_specifiers`. Subclassing
it registers the wrapper as a field specifier, so the false
positives disappear with no runtime change. Move the `field` helper
next to `Module` in `_jaxext` (a foundational, equinox-free,
bartz-internal layer) and reparent the dataclasses that use it:
grove `HeapArrays` (hence `Forest`/`TreesTrace`), mcmcstep `State`,
`StepConfig`, `Forest` and the `_step` helpers, `mcmcloop._trace`,
and `debug._traceconv`.

Notes:
- the metaclass is derived via `type(Module)`, not equinox's private
  metaclass, whose name differs across versions (`_ActualModuleMeta`
  on 0.12.2 vs `_ModuleMeta` on 0.13.8);
- `dataclass_transform` is imported under a `python<3.11` guard;
- `field` keeps no return annotation on purpose: `-> Field[Any]`
  makes ty model a defaulted field's value as `Field[Field[Any]]`
  and reject it against the declared array type.
