+++ src/bartz/mcmcstep/_state.py
@@ -32,7 +32,7 @@ from typing import Literal, TypedDict, TypeVar
 
 import jax
 import numpy
-from equinox import Module, error_if, filter_jit
+from equinox import error_if, filter_jit
 from jax import NamedSharding, device_put, lax, make_mesh, random, shard_map, tree, vmap
 from jax import numpy as jnp
 from jax.scipy.linalg import solve_triangular
@@ -40,9 +40,9 @@ from jax.sharding import AxisType, Mesh, PartitionSpec
 from jaxtyping import Array, Bool, Float, Float32, Int32, Integer, Key, PyTree, UInt
 from numpy import ndarray
 
-from bartz._jaxext import jaxtyping_disabled, jit, minimal_unsigned_dtype
+from bartz._jaxext import Module, field, jaxtyping_disabled, jit, minimal_unsigned_dtype
 from bartz.grove import HeapArrays, tree_depths
-from bartz.mcmcstep._axes import CHAIN_AXIS, chain_vmap_axes, data_vmap_axes, field
+from bartz.mcmcstep._axes import CHAIN_AXIS, chain_vmap_axes, data_vmap_axes
 from bartz.mcmcstep._lazy import (
     _is_lazy_or_none,
     _lazy_from_array,
